requirements.txt
altgraph==0.15
astroid==1.6.2
coverage==5.3
future==0.16.0
isort==4.3.4
lazy-object-proxy==1.3.1
macholib==1.9
mccabe==0.6.1
pefile==2017.11.5
pylint==1.8.3
pyperclip==1.8.1
python-dotenv==0.8.2
PyYAML==5.1
selenium-base==1.51.0
six==1.11.0
wrapt==1.10.11
